backend/requirements.txt
fastapi==0.118.0
jose==1.0.0
pydantic==2.11.9
python-dotenv==1.1.1
python_bcrypt==0.3.2
sqlmodel==0.0.25
